fix: 깃 오류 수정
